# lab2_new/movie-service/src/main/java/com/lab2/movie/model/MovieListResponse.java
package com.lab2.movie.model;

import javax.xml.bind.annotation.XmlElement;
import javax.xml.bind.annotation.XmlRootElement;
import java.util.List;

@XmlRootElement(name = "movieListResponse")
public class MovieListResponse {
    private List<Movie> movies;
    private long totalElements;
    private int totalPages;
    private int currentPage;
    private int pageSize;

    public MovieListResponse() {}

    public MovieListResponse(List<Movie> movies, long totalElements, int totalPages, int currentPage, int pageSize) {
        this.movies = movies;
        this.totalElements = totalElements;
        this.totalPages = totalPages;
        this.currentPage = currentPage;
        this.pageSize = pageSize;
    }

    @XmlElement
    public List<Movie> getMovies() {
        return movies;
    }

    public void setMovies(List<Movie> movies) {
        this.movies = movies;
    }

    @XmlElement
    public long getTotalElements() {
        return totalElements;
    }

    public void setTotalElements(long totalElements) {
        this.totalElements = totalElements;
    }

    @XmlElement
    public int getTotalPages() {
        return totalPages;
    }

    public void setTotalPages(int totalPages) {
        this.totalPages = totalPages;
    }

    @XmlElement
    public int getCurrentPage() {
        return currentPage;
    }

    public void setCurrentPage(int currentPage) {
        this.currentPage = currentPage;
    }

    @XmlElement
    public int getPageSize() {
        return pageSize;
    }

    public void setPageSize(int pageSize) {
        this.pageSize = pageSize;
    }
}